backend/requirements.txt
fastapi
sqlalchemy==1.4.41
typing-extensions==4.3.0; python_version < '3.11'
pydantic==1.10.3
psycopg2-binary==2.9.5
pyJWT==2.6.0
passlib==1.7.4
python-multipart==0.0.5
uvicorn==0.20.0
bcrypt==4.0.1
pandas==1.5.3
pmdarima==2.0.2